skip null when computing numerical_hsitogram

+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/TypeCoercion.scala
@@ -347,11 +347,11 @@ object TypeCoercion {
       case VariancePop(e @ StringType()) => VariancePop(Cast(e, DoubleType))
       case VarianceSamp(e @ StringType()) => VarianceSamp(Cast(e, DoubleType))
       case NumericHistogram(e @ StringType(), b @ StringType()) =>
-        NumericHistogram(Cast(e, DoubleType), Cast(b, DoubleType))
+        NumericHistogram(Cast(e, DoubleType), Cast(b, IntegerType))
       case NumericHistogram(e @ NumericType(), b @ StringType()) =>
-        NumericHistogram(Cast(e, DoubleType), Cast(b, DoubleType))
+        NumericHistogram(Cast(e, DoubleType), Cast(b, IntegerType))
       case NumericHistogram(e @ StringType(), b @ NumericType()) =>
-        NumericHistogram(Cast(e, DoubleType), Cast(b, DoubleType))
+        NumericHistogram(Cast(e, DoubleType), Cast(b, IntegerType))
       case Skewness(e @ StringType()) => Skewness(Cast(e, DoubleType))
       case Kurtosis(e @ StringType()) => Kurtosis(Cast(e, DoubleType))
     }
